Refactor move fish logic
